FMOD Implementation and AudioManager Creation
FMOD's files has been added, disabling Unity's built in audio and adding to the gitignore for its files.

AudioManager implemented for usage to change Master, Music, and SFX channels of volume. As well, handling transitions in music for different areas.

The AreaMusicTrigger monobehavior can notify the AudioManager of area on trigger with a collider.